perf(tests): split publicApiContext2 out of createContext
`publicApiContext2` is the only thing in createContext that needs `appRouter`, so
importing the module for the `Context` type — which ~100 controllers do — put the
entire tRPC router in the graph. It also made createContext part of a
createContext -> routers -> controllers -> createContext cycle.

Function body relocated verbatim. The six v1 routes that call it import the new
module; nothing else in createContext changes.

Split out of #3958.
