+++ src/server/createContext.ts
@@ -4,8 +4,6 @@ import { Tracker } from './clickhouse/client';
 import { resolveClientIpOrNull } from '~/server/utils/client-ip';
 import { isProd } from '~/env/other';
 import { getFeatureFlagsLazy } from '~/server/services/feature-flags.service';
-import { createCallerFactory } from '~/server/trpc';
-import { appRouter } from '~/server/routers';
 import { getRequestDomainColor } from '~/server/utils/server-domain';
 import { isAllowedOriginRequest } from '~/server/utils/origin-helpers';
 import { TokenScope } from '~/shared/constants/token-scope.constants';
@@ -107,36 +105,6 @@ export const createContext = async ({
     apiKeyId,
     subject,
   };
-};
-
-const createCaller = createCallerFactory(appRouter);
-export const publicApiContext2 = async (req: NextApiRequest, res: NextApiResponse) => {
-  const domain = getRequestDomainColor(req) ?? 'blue';
-
-  return createCaller({
-    user: undefined,
-    acceptableOrigin: true,
-    features: getFeatureFlagsLazy({ req }),
-    track: new Tracker(req, res),
-    // ATTRIBUTION surface — see the note on `ip` in `createContext` above.
-    ip: resolveClientIpOrNull(req) ?? '',
-    cache: {
-      browserTTL: 3 * 60,
-      edgeTTL: 3 * 60,
-      staleWhileRevalidate: 60,
-      canCache: true,
-      skip: false,
-    },
-    res,
-    req,
-    domain,
-    // Non-client-facing context â€” use an always-open signal so downstream
-    // callers that expect AbortSignal have a valid value.
-    signal: new AbortController().signal,
-    tokenScope: TokenScope.Full,
-    apiKeyId: undefined,
-    subject: undefined,
-  });
 };
 
 export const publicApiContext = async (req: NextApiRequest, res: NextApiResponse) => {
