+++ src/tests/api/rest-envelope-consolidation.test.ts
@@ -73,7 +73,7 @@ const {
   mockTracker: vi.fn(),
 }));
 
-vi.mock('~/server/createContext', async (importOriginal) => ({
+vi.mock('~/server/public-api-context', async (importOriginal) => ({
   ...(await importOriginal<Record<string, unknown>>()),
   publicApiContext2: mockPublicApiContext2,
 }));
