+++ src/tests/api/v1/creators.test.ts
@@ -12,7 +12,7 @@ const { mockPublicApiContext2, mockGetCreators } = vi.hoisted(() => ({
   mockGetCreators: vi.fn(),
 }));
 
-vi.mock('~/server/createContext', () => ({
+vi.mock('~/server/public-api-context', () => ({
   publicApiContext2: mockPublicApiContext2,
 }));
