+++ src/tests/api/v1/tags.test.ts
@@ -11,7 +11,7 @@ const { mockPublicApiContext2, mockGetAll } = vi.hoisted(() => ({
   mockGetAll: vi.fn(),
 }));
 
-vi.mock('~/server/createContext', () => ({
+vi.mock('~/server/public-api-context', () => ({
   publicApiContext2: mockPublicApiContext2,
 }));
